+++ src/main/java/com/girigiri/kwrental/reservation/service/remainquantity/RemainingQuantityServiceImpl.java
@@ -1,4 +1,4 @@
-package com.girigiri.kwrental.reservation.service;
+package com.girigiri.kwrental.reservation.service.remainquantity;
 
 import static java.util.stream.Collectors.*;
 
@@ -11,11 +11,8 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.girigiri.kwrental.asset.domain.Rentable;
-import com.girigiri.kwrental.asset.service.AssetService;
 import com.girigiri.kwrental.asset.service.RemainingQuantityService;
 import com.girigiri.kwrental.reservation.domain.OperatingPeriod;
-import com.girigiri.kwrental.reservation.domain.RentalPeriod;
 import com.girigiri.kwrental.reservation.domain.ReservationSpec;
 import com.girigiri.kwrental.reservation.domain.ReservedAmount;
 import com.girigiri.kwrental.reservation.repository.ReservationSpecRepository;
@@ -24,10 +21,9 @@ import lombok.RequiredArgsConstructor;
 
 @Service
 @RequiredArgsConstructor
-public class RemainingQuantityServiceImpl implements RemainingQuantityService, AmountValidator {
+public class RemainingQuantityServiceImpl implements RemainingQuantityService {
 
 	private final ReservationSpecRepository reservationSpecRepository;
-	private final AssetService assetService;
 
 	@Override
 	@Transactional(readOnly = true, propagation = Propagation.MANDATORY)
@@ -36,26 +32,6 @@ public class RemainingQuantityServiceImpl implements RemainingQuantityService, A
 		return reservationSpecRepository.findRentalAmountsByAssetIds(rentableIds, date)
 			.stream()
 			.collect(toMap(ReservedAmount::getEquipmentId, ReservedAmount::getRemainingAmount));
-	}
-
-	@Override   // TODO: 2023/04/23 반복문을 두번 도는 로직을 최적화 할 수 있다.
-	@Transactional(readOnly = true, propagation = Propagation.MANDATORY)
-	public void validateAmount(final Long assetId, final Integer amount, final RentalPeriod rentalPeriod) {
-		final Rentable rentable = assetService.getRentableById(assetId);
-		final List<ReservationSpec> overlappedReservationSpecs = reservationSpecRepository.findOverlappedReservedOrRentedByPeriod(
-			assetId, rentalPeriod);
-		for (LocalDate i = rentalPeriod.getRentalStartDate(); i.isBefore(
-			rentalPeriod.getRentalEndDate()); i = i.plusDays(1)) {
-			final int rentedAmountByDate = sumRentedAmountByDate(overlappedReservationSpecs, i);
-			rentable.validateAmountForRent(amount + rentedAmountByDate);
-		}
-	}
-
-	private int sumRentedAmountByDate(final List<ReservationSpec> overlappedReservationSpecs, final LocalDate date) {
-		return overlappedReservationSpecs.stream()
-			.filter(spec -> spec.containsDate(date))
-			.mapToInt(spec -> spec.getAmount().getAmount())
-			.sum();
 	}
 
 	@Override
