+++ src/test/java/com/girigiri/kwrental/reservation/service/remainquantity/RemainingQuantityServiceImplTest.java
@@ -1,6 +1,5 @@
-package com.girigiri.kwrental.reservation.service;
+package com.girigiri.kwrental.reservation.service.remainquantity;
 
-import static org.assertj.core.api.AssertionsForClassTypes.*;
 import static org.mockito.ArgumentMatchers.*;
 import static org.mockito.BDDMockito.*;
 
@@ -17,11 +16,9 @@ import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 
 import com.girigiri.kwrental.asset.equipment.domain.Equipment;
-import com.girigiri.kwrental.asset.service.AssetService;
 import com.girigiri.kwrental.reservation.domain.RentalAmount;
 import com.girigiri.kwrental.reservation.domain.RentalPeriod;
 import com.girigiri.kwrental.reservation.domain.ReservationSpec;
-import com.girigiri.kwrental.reservation.exception.NotEnoughAmountException;
 import com.girigiri.kwrental.reservation.repository.ReservationSpecRepository;
 import com.girigiri.kwrental.testsupport.fixture.EquipmentFixture;
 import com.girigiri.kwrental.testsupport.fixture.ReservationSpecFixture;
@@ -32,59 +29,8 @@ class RemainingQuantityServiceImplTest {
     @Mock
     private ReservationSpecRepository reservationSpecRepository;
 
-    @Mock
-    private AssetService assetService;
-
     @InjectMocks
     private RemainingQuantityServiceImpl remainingQuantityService;
-
-    @Test
-    @DisplayName("대여 가능한 갯수를 확인")
-    void validateAmount() {
-        // given
-        final Equipment equipment = EquipmentFixture.builder().id(1L).totalQuantity(2).build();
-        LocalDate now = LocalDate.now();
-        final ReservationSpec reservationSpec1 = ReservationSpecFixture.builder(equipment)
-            .amount(RentalAmount.ofPositive(1))
-            .period(new RentalPeriod(now, now.plusDays(1)))
-            .build();
-        final ReservationSpec reservationSpec2 = ReservationSpecFixture.builder(equipment)
-            .amount(RentalAmount.ofPositive(1))
-            .period(new RentalPeriod(now.plusDays(1), now.plusDays(2)))
-            .build();
-
-        given(assetService.getRentableById(any())).willReturn(equipment);
-        given(reservationSpecRepository.findOverlappedReservedOrRentedByPeriod(any(), any())).willReturn(
-            List.of(reservationSpec1, reservationSpec2));
-
-        // when, then
-        assertThatCode(() -> remainingQuantityService.validateAmount(1L, 1, new RentalPeriod(now, now.plusDays(2))))
-            .doesNotThrowAnyException();
-    }
-
-    @Test
-    @DisplayName("대여 가능한 갯수가 요청한 갯수보다 부족하면 예외 발생")
-    void validateAmount_notEnoughRemainingAmount() {
-        // given
-        final Equipment equipment = EquipmentFixture.builder().id(1L).totalQuantity(2).build();
-        LocalDate now = LocalDate.now();
-        final ReservationSpec reservationSpec1 = ReservationSpecFixture.builder(equipment)
-            .amount(RentalAmount.ofPositive(1))
-            .period(new RentalPeriod(now, now.plusDays(1)))
-            .build();
-        final ReservationSpec reservationSpec2 = ReservationSpecFixture.builder(equipment)
-            .amount(RentalAmount.ofPositive(1))
-            .period(new RentalPeriod(now, now.plusDays(1)))
-            .build();
-
-        given(assetService.getRentableById(any())).willReturn(equipment);
-        given(reservationSpecRepository.findOverlappedReservedOrRentedByPeriod(any(), any())).willReturn(
-            List.of(reservationSpec1, reservationSpec2));
-
-        // when, then
-        assertThatThrownBy(() -> remainingQuantityService.validateAmount(1L, 1, new RentalPeriod(now, now.plusDays(1))))
-            .isExactlyInstanceOf(NotEnoughAmountException.class);
-    }
 
     @Test
     @DisplayName("기간 사이의 대여 신청 대여 갯수를 센다.")
